Initial isolation of flag checks. Diff adjustment still done in side effects
