+++ vessel/utils/diffoscope.py
@@ -29,12 +29,10 @@ import re
 from pathlib import Path
 from typing import Any, Optional
 
-import magic
-
-from vessel.diff.helpers.diffline import DiffLine
 from vessel.diff.helpers.failure import Failure
 from vessel.diff.helpers.flag import Flag
 from vessel.diff.helpers.file_diff import FileDiff
+from vessel.utils.flag_check import check_flags
 from vessel.utils.unified_diff import (
     failures_from_difflines,
     intervals_to_str,
@@ -226,7 +224,7 @@ class DiffoscopeParser:
         if parent_comments:
             temp_comments.extend(parent_comments)
 
-        diff = FileDiff(
+        file_diff = FileDiff(
             detail["source1"],
             detail["source2"],
             temp_comments,
@@ -239,25 +237,28 @@ class DiffoscopeParser:
             not Path(detail["source1"]).is_absolute()
             or not Path(detail["source2"]).is_absolute()
         ):
-            diff.command = detail["source1"]
-            diff.source1 = parent_source1
-            diff.source2 = parent_source2
+            file_diff.command = detail["source1"]
+            file_diff.source1 = parent_source1
+            file_diff.source2 = parent_source2
 
         is_binary = bool(detail.get("has_internal_linenos"))
         for minus_line, plus_line in zip(
-            diff.minus_aligned_lines,
-            diff.plus_aligned_lines,
+            file_diff.minus_aligned_lines,
+            file_diff.plus_aligned_lines,
             strict=False,
         ):
-            self._check_flags(diff, minus_line, plus_line, is_binary)
+            failure_summary = check_flags(self.flags, self.filetype_lookup1, self.filetype_lookup2, file_diff, minus_line, plus_line, is_binary)
+            self.trivial_failure_count += failure_summary.trivial_failures
+            self.nontrivial_failure_count += failure_summary.nontrivial_failures
+            self.unknown_failure_count += failure_summary.unknown_failures
 
             # Check so line by line comparison don't happen in binary diffs and
             # this is after all the flags have been checked so the diff is done
             # being evaluated
             if is_binary:
-                if len(diff.flagged_failures) == 0:
+                if len(file_diff.flagged_failures) == 0:
                     self.unknown_failure_count += 1
-                    diff.unknown_failures.append(
+                    file_diff.unknown_failures.append(
                         {
                             "comments": [
                                 "Flag indiff regex are not ran on binary "
@@ -287,7 +288,7 @@ class DiffoscopeParser:
             )
             if minus_unmatched_str != plus_unmatched_str:
                 self.unknown_failure_count += 1
-                diff.unknown_failures.append(
+                file_diff.unknown_failures.append(
                     Failure(
                         minus_line if minus_line else None,
                         plus_line if plus_line else None,
@@ -296,182 +297,4 @@ class DiffoscopeParser:
                     ).to_dict(),
                 )
 
-        self.diff_list.append(diff.to_dict())
-
-    def _check_flags(
-        self: "DiffoscopeParser",
-        diff: FileDiff,
-        minus_line: DiffLine,
-        plus_line: DiffLine,
-        is_binary: bool,
-    ):
-        """Check a Diff against all flags and update Diff based on matches or non matches.
-
-        Take in a Diff, iterate through all of the flags and check if each matches the difference and
-        the minus and plus lines of the Diff while updating the failure counts, and the lists of
-        failures in the Diff parameter object.
-
-        Args:
-            diff: Diff object to be checked
-            minus_line: Line of the minus file in the unified diff to be checked
-            plus_line: Line of the plus file in the unified diff to be checked
-        """
-        for flag in self.flags:
-            flag_matches = True
-
-            # Check if filepath matches flag
-            flag_matches = self._check_flag_filepath(
-                flag, diff.source1, diff.source2
-            )
-
-            # Check if filetype matches flag
-            if flag_matches:
-                flag_matches = self._check_flag_filetype(
-                    flag, diff.source1, diff.source2
-                )
-
-            # Check if command matches flag
-            if flag_matches:
-                flag_matches = self._check_flag_command(flag, diff.command)
-
-            # Check if comment matches flag
-            if flag_matches:
-                flag_matches = self._check_flag_comment(flag, diff.comments)
-
-            # Handle a binary line that matches the flag
-            if (
-                flag_matches
-                and is_binary
-                and flag.regex["indiff"] == re.compile(".*")
-            ):
-                diff.flagged_failures.append(
-                    {
-                        "id": flag.flag_id,
-                        "description": flag.description,
-                        "metadata": getattr(flag, "metadata", False),
-                        "comments": [
-                            "Flag indiff regex are not ran on binary "
-                            "unified diff. However this matched all "
-                            "of the other criteria for this flag.",
-                        ],
-                    },
-                )
-            # Handle any non-binary line that matches the flag
-            elif flag_matches:
-                (
-                    flagged_failure_list,
-                    unknown_failure_list,
-                    minus_line.unmatched_intervals,
-                    plus_line.unmatched_intervals,
-                ) = failures_from_difflines(
-                    minus_line,
-                    plus_line,
-                    flag,
-                )
-                # Check to not create duplicate matches on flags that match based on filepath, filetype, command or comment
-                #     and have indiff set to ".*"
-                if flag.regex["indiff"] != re.compile(
-                    ".*"
-                ) or flag.flag_id not in [
-                    flag["id"] for flag in diff.flagged_failures
-                ]:
-                    for failure in flagged_failure_list:
-                        failure["metadata"] = flag.metadata
-                        failure["severity"] = flag.severity
-                        if flag.severity == "Low":
-                            self.trivial_failure_count += 1
-                        else:
-                            self.nontrivial_failure_count += 1
-                    self.unknown_failure_count += len(unknown_failure_list)
-                    diff.flagged_failures.extend(flagged_failure_list)
-                    diff.unknown_failures.extend(unknown_failure_list)
-
-    def _check_flag_filepath(
-        self: "DiffoscopeParser", flag: Flag, source1: str, source2: str
-    ) -> bool:
-        """Check difference sources against filepath regex of flag.
-
-        Args:
-            flag: Flag to check against
-            source1: String representing filepath to source1
-            source2: String representing filepath to source2
-        """
-        if not flag.regex["filepath"].search(source1) or not flag.regex[
-            "filepath"
-        ].search(source2):
-            return False
-        else:
-            return True
-
-    def _check_flag_filetype(
-        self: "DiffoscopeParser", flag: Flag, source1: str, source2: str
-    ) -> bool:
-        """Check difference sources against filetype regex of flag.
-
-        Perform check of source filetypes. If both files exist locally, use
-        magic library for data type otherwise use types from the metadata.
-
-        Args:
-            flag: Flag to check against
-            source1: String representing filepath to source1
-            source2: String representing filepath to source2
-        """
-        source_1_exists = Path(source1).is_file()
-        source_2_exists = Path(source2).is_file()
-        if source_1_exists and source_2_exists:
-            file_type_1 = magic.from_file(source1)
-            file_type_2 = magic.from_file(source2)
-            if not flag.regex["filetype"].search(
-                file_type_1
-            ) or not flag.regex["filetype"].search(file_type_2):
-                return False
-        else:
-            # Local file does not exist: try checksum metadata lookups.
-            if (
-                self.filetype_lookup1 is not None
-                and self.filetype_lookup2 is not None
-            ):
-                file_type_1 = self.filetype_lookup1.get(source1, "")
-                file_type_2 = self.filetype_lookup2.get(source2, "")
-
-                if file_type_1 and file_type_2:
-                    if not flag.regex["filetype"].search(
-                        file_type_1
-                    ) or not flag.regex["filetype"].search(file_type_2):
-                        return False
-
-        return True
-
-    def _check_flag_command(
-        self: "DiffoscopeParser", flag: Flag, command: str
-    ) -> bool:
-        """Check difference command against command regex of flag.
-
-        Command of the difference will be populated if the diff was found by diffoscope
-        using a command such as stat {}.
-
-        Args:
-            flag: Flag to check against
-            command: Command used to find diff
-        """
-        if not flag.regex["command"].search(command):
-            return False
-        else:
-            return True
-
-    def _check_flag_comment(
-        self: "DiffoscopeParser", flag: Flag, comments: list
-    ) -> bool:
-        """Check difference comments against comment regex of flag.
-
-        Checks if any comment of the command matches, or if the comment list is
-        empty and the regex is set to accept any value.
-        """
-        if comments != [] and not any(
-            flag.regex["comment"].search(comment) for comment in comments
-        ):
-            return False
-        elif comments == [] and flag.regex["comment"] != re.compile(".*"):
-            return False
-        else:
-            return True
+        self.diff_list.append(file_diff.to_dict())
